Flyttet VelgVedtaksresultat og VedtaksresultatSwitch til egne komponenter. Oppdatert med håndtering av dobbeltklikk
