+++ src/frontend/komponenter/Behandling/VedtakOgBeregning/VedtaksresultatSwitch.tsx
@@ -1,0 +1,135 @@
+import React, { Dispatch, SetStateAction, useState } from 'react';
+import { EBehandlingResultat, IVedtak } from '../../../typer/vedtak';
+import { Element } from 'nav-frontend-typografi';
+import { Textarea } from 'nav-frontend-skjema';
+import { Hovedknapp } from 'nav-frontend-knapper';
+import { Ressurs, RessursStatus } from '../../../typer/ressurs';
+import { useApp } from '../../../context/AppContext';
+import { ModalAction, ModalType, useModal } from '../../../context/ModalContext';
+import styled from 'styled-components';
+import { AlertStripeAdvarsel } from 'nav-frontend-alertstriper';
+import { useHistory } from 'react-router-dom';
+
+interface Props {
+    vedtaksresultatType: EBehandlingResultat;
+    behandlingId: string;
+    settFeilmelding: Dispatch<SetStateAction<string>>;
+}
+
+const StyledAdvarsel = styled(AlertStripeAdvarsel)`
+    margin-top: 2rem;
+`;
+
+const VedtaksresultatSwitch: React.FC<Props> = (props: Props) => {
+    const { axiosRequest } = useApp();
+    const { modalDispatch } = useModal();
+    const history = useHistory();
+    const [periodeBegrunnelse, settPeriodeBegrunnelse] = useState<string>('');
+    const [inntektBegrunnelse, settInntektBegrunnelse] = useState<string>('');
+    const [laster, settLaster] = useState<boolean>(false);
+    const { vedtaksresultatType, behandlingId, settFeilmelding } = props;
+
+    const lagBlankett = () => {
+        settLaster(true);
+        axiosRequest<string, IVedtak>({
+            method: 'POST',
+            url: `/familie-ef-sak/api/beregning/${behandlingId}/lagre-vedtak`,
+            data: {
+                resultatType: vedtaksresultatType,
+                periodeBegrunnelse,
+                inntektBegrunnelse,
+            },
+        })
+            .then((res: Ressurs<string>) => {
+                switch (res.status) {
+                    case RessursStatus.SUKSESS:
+                        history.push(`/behandling/${behandlingId}/blankett`);
+                        break;
+                    case RessursStatus.HENTER:
+                    case RessursStatus.IKKE_HENTET:
+                        break;
+                    default:
+                        settFeilmelding(res.frontendFeilmelding);
+                }
+            })
+            .finally(() => {
+                settLaster(false);
+            });
+    };
+
+    const behandleIGosys = () => {
+        settLaster(true);
+        axiosRequest<{ id: string }, null>({
+            method: 'POST',
+            url: `/familie-ef-sak/api/behandling/${behandlingId}/annuller`,
+        })
+            .then((res: Ressurs<{ id: string }>) => {
+                switch (res.status) {
+                    case RessursStatus.SUKSESS:
+                        modalDispatch({
+                            type: ModalAction.VIS_MODAL,
+                            modalType: ModalType.BEHANDLES_I_GOSYS,
+                        });
+                        break;
+                    case RessursStatus.HENTER:
+                    case RessursStatus.IKKE_HENTET:
+                        break;
+                    default:
+                        settFeilmelding(res.frontendFeilmelding);
+                }
+            })
+            .finally(() => {
+                settLaster(false);
+            });
+    };
+
+    switch (vedtaksresultatType) {
+        case EBehandlingResultat.INNVILGE:
+            return (
+                <>
+                    <Element style={{ marginBottom: '1rem', marginTop: '3rem' }}>
+                        Vedtaksperiode
+                    </Element>
+                    <Textarea
+                        value={periodeBegrunnelse}
+                        onChange={(e) => {
+                            settPeriodeBegrunnelse(e.target.value);
+                        }}
+                        label="Begrunnelse"
+                    />
+                    <Element style={{ marginBottom: '1rem', marginTop: '3rem' }}>Inntekt</Element>
+                    <Textarea
+                        value={inntektBegrunnelse}
+                        onChange={(e) => {
+                            settInntektBegrunnelse(e.target.value);
+                        }}
+                        label="Begrunnelse"
+                    />
+                    <Hovedknapp
+                        style={{ marginTop: '2rem' }}
+                        onClick={lagBlankett}
+                        disabled={laster}
+                    >
+                        Lag blankett
+                    </Hovedknapp>
+                </>
+            );
+        case EBehandlingResultat.BEHANDLE_I_GOSYS:
+            return (
+                <>
+                    <StyledAdvarsel>Oppgaven annulleres og må fullføres i Gosys</StyledAdvarsel>
+                    <Hovedknapp
+                        style={{ marginTop: '2rem' }}
+                        onClick={behandleIGosys}
+                        disabled={laster}
+                    >
+                        Avslutt og behandle i Gosys
+                    </Hovedknapp>
+                </>
+            );
+        default:
+            return null;
+    }
+};
+
+export default VedtaksresultatSwitch;
